+++ src/Aardvark.Dom/SceneGraph/SceneHandler.fs
@@ -1043,7 +1043,7 @@ type SceneHandler(signature : IFramebufferSignature, trigger : RenderControlEven
 
         capturedResult
 
-    member x.HandlePointerEvent(kind : SceneEventKind, pixel : V2i, ctrl : bool, shift : bool, alt : bool, meta : bool, pointerId : int, ?scrollDelta : V2d, ?button : int) : bool =
+    member x.HandlePointerEvent(kind : SceneEventKind, pixel : V2i, ctrl : bool, shift : bool, alt : bool, meta : bool, pointerId : int, ?scrollDelta : V2d, ?button : Button) : bool =
         let s = viewportSize
         if s.AllGreater 0 then
             let view = AVal.force view 
@@ -1053,7 +1053,7 @@ type SceneHandler(signature : IFramebufferSignature, trigger : RenderControlEven
             match x.Read(pixel, pointerId) with
             | Some (best, target, depth, viewNormal) ->
                 let model = TraversalState.modelTrafo best
-                let button = defaultArg button -1
+                let button = defaultArg button Button.None
                 let loc = SceneEventLocation(model, view, proj, V2d pixel, s, depth, viewNormal)
                 let evt = ScenePointerEvent(x, best, target, kind, loc, ctrl, shift, alt, meta, scrollDelta, pointerId, button)
 
@@ -1080,7 +1080,7 @@ type SceneHandler(signature : IFramebufferSignature, trigger : RenderControlEven
             | None ->
                 match kind with
                 | SceneEventKind.Click when Option.isSome lastFocus.Value -> 
-                    let button = defaultArg button -1
+                    let button = defaultArg button Button.None
                     let loc = SceneEventLocation(AVal.constant Trafo3d.Identity, view, proj, V2d pixel, s, 1.0, V3d.Zero)
                     let evt = ScenePointerEvent(x, lastFocus.Value.Value, null, kind, loc, ctrl, shift, alt, meta, scrollDelta, pointerId, button)
                     TraversalState.handleDifferential lastFocus SceneEventKind.FocusEnter SceneEventKind.FocusLeave evt None
@@ -1088,7 +1088,7 @@ type SceneHandler(signature : IFramebufferSignature, trigger : RenderControlEven
                     
 
                 if Option.isSome lastOver.Value then
-                    let button = defaultArg button -1
+                    let button = defaultArg button Button.None
                     let loc = SceneEventLocation(AVal.constant Trafo3d.Identity, view, proj, V2d pixel, s, 1.0, V3d.Zero)
                     let evt = ScenePointerEvent(x, lastOver.Value.Value, null, kind, loc, ctrl, shift, alt, meta, scrollDelta, pointerId, button)
                     TraversalState.handleMove lastOver evt None
@@ -1121,7 +1121,7 @@ type SceneHandler(signature : IFramebufferSignature, trigger : RenderControlEven
                 | Some s -> s :> obj
                 | None -> null
 
-            let evt = ScenePointerEvent(x, target, target, SceneEventKind.FocusLeave, evtLocation, false, false, false, false, V2d.Zero, -1, -1)
+            let evt = ScenePointerEvent(x, target, target, SceneEventKind.FocusLeave, evtLocation, false, false, false, false, V2d.Zero, -1, Button.None)
             TraversalState.handleDifferential lastFocus SceneEventKind.FocusEnter SceneEventKind.FocusLeave evt newTarget
                 
     member x.HandleKeyEvent(kind : SceneEventKind, ctrl : bool, shift : bool, alt : bool, meta : bool, code : string, key : string, keyLocation : KeyLocation, text : string, isRepeat : bool) : bool =
@@ -1152,7 +1152,7 @@ type SceneHandler(signature : IFramebufferSignature, trigger : RenderControlEven
             | Some (:? TraversalState as s) -> x.SetFocus(Some s)
             | _ -> x.SetFocus None
             
-        member x.HandlePointerEvent(kind : SceneEventKind, pixel : V2i, ctrl : bool, shift : bool, alt : bool, meta : bool, scrollDelta : V2d, pointerId : int, button : int) =
+        member x.HandlePointerEvent(kind : SceneEventKind, pixel : V2i, ctrl : bool, shift : bool, alt : bool, meta : bool, scrollDelta : V2d, pointerId : int, button : Button) =
             x.HandlePointerEvent(kind, pixel, ctrl, shift, alt, meta, pointerId, scrollDelta, button)
               
         member x.HandleKeyEvent(kind : SceneEventKind, ctrl : bool, shift : bool, alt : bool, meta : bool, code : string, key : string, keyLocation : KeyLocation, text : string, isRepeat : bool) =
@@ -1183,7 +1183,7 @@ type SceneHandler(signature : IFramebufferSignature, trigger : RenderControlEven
                     | Some t -> t :> obj
                     | None -> null
 
-                let evt = ScenePointerEvent(x, eventTarget, eventTarget, SceneEventKind.PointerMove, loc, false, false, false, false, V2d.Zero, pointerId, -1)
+                let evt = ScenePointerEvent(x, eventTarget, eventTarget, SceneEventKind.PointerMove, loc, false, false, false, false, V2d.Zero, pointerId, Button.None)
                 TraversalState.handleMove lastOver evt target
                 match target with
                 | Some t -> TraversalState.handleEvent true evt t |> ignore