+++ RELEASE_NOTES.md
@@ -1,3 +1,8 @@
+### 1.0.0-prerelease0009
+* Sg PointerEvents get Button instead of int
+* Sg.Text no longer uses Aardvark.SceneGraph
+* added Sg.Shape
+
 ### 1.0.0-prerelease0008
 * Sg.Cursor now takes a string
 * Dom KeyboardEvents and InputEvents